requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
docker>=6.1.0
gitpython>=3.1.40
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
loguru>=0.7.0
# File handling
pyyaml>=6.0.1
python-docx>=0.8.11
PyPDF2>=3.0.0
pandas>=2.1.1
# Monitoring and Dashboard
prometheus-client>=0.17.1
streamlit>=1.24.0
plotly>=5.17.0
psutil>=5.9.0
